Move auth tests to auth package
